--- src/test/java/read.me
@@ -1,1 +1,0 @@
-
